feat: move nav to component
