+++ client/dashboard/src/pages/logs/Logs.tsx
@@ -1,45 +1,9 @@
-import { Link, Outlet, useLocation } from "react-router";
+import { Outlet } from "react-router";
 import { LogsContent } from "@/components/observe/LogsContent";
 import { Page } from "@/components/page-layout";
 import { RequireScope } from "@/components/require-scope";
-import { useSlugs } from "@/contexts/Sdk";
-import { cn } from "@/lib/utils";
 import { LogsHooks } from "@/components/observe/LogsHooks";
-
-function LogsTabNav() {
-  const { orgSlug, projectSlug } = useSlugs();
-  const location = useLocation();
-
-  const base = `/${orgSlug}/${projectSlug}/logs`;
-  const tabs = [
-    { label: "Hooks", href: `${base}/hooks` },
-    { label: "MCP Servers", href: `${base}/mcp` },
-    { label: "Agents", href: `${base}/agents` },
-  ];
-
-  return (
-    <div className="border-border flex h-auto w-full items-end gap-8 border-b bg-transparent px-8 pt-2">
-      {tabs.map((tab) => {
-        const isActive = location.pathname === tab.href;
-        return (
-          <Link
-            key={tab.href}
-            to={tab.href}
-            className={cn(
-              "relative flex-none pb-3 text-sm font-medium no-underline transition-colors",
-              "after:absolute after:right-0 after:bottom-0 after:left-0 after:h-0.5",
-              isActive
-                ? "text-foreground after:bg-primary"
-                : "text-muted-foreground hover:text-foreground after:bg-transparent",
-            )}
-          >
-            {tab.label}
-          </Link>
-        );
-      })}
-    </div>
-  );
-}
+import { ObserveTabNav } from "@/components/observe/ObserveTabNav";
 
 export function LogsRoot() {
   return (
@@ -47,7 +11,7 @@ export function LogsRoot() {
       <Page.Header>
         <Page.Header.Breadcrumbs fullWidth />
       </Page.Header>
-      <LogsTabNav />
+      <ObserveTabNav base="logs" />
       <Page.Body fullWidth noPadding>
         <Outlet />
       </Page.Body>
